Encode strings directly into AST

+++ src/vm.h
@@ -8,7 +8,7 @@ struct v7_ast;    /* V7 Abstract Syntax Tree. */
 struct v7_arg;    /* C/JavaScript function parameters */
 
 typedef double v7_num_t;    /* Override to integer on systems with no MMU */
-typedef unsigned long v7_strlen_t;
+typedef unsigned int v7_strlen_t;
 typedef void (*v7_func2_t)(struct v7_arg *arg);
 
 /*
